pyproject.toml
[tool.poetry]
name = "poetic-emr-serverless"
version = "0.1.0"
description = ""
authors = ["kazdy <d.kazmirski@gmail.com>"]
readme = "README.md"
packages = [{include = "poetic_emr_serverless"}]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
cloudpathlib = "^0.10.0"

[tool.poetry.group.dev.dependencies]
pyspark = "3.3.0"
pytest = "^7.2.0"
chispa = "^0.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src", "tests"]
testpaths = ["tests"]
